* wmiatlprov.h (CInstanceProviderHelper::CheckInstancePath): 	Cast first parameter in call to GetInfo to ULONG. 	* winioctl.h (IsDsmActionNonDestructive): Replace test for != 0L 	with logical negation operator. 	* wininet.h (INTERNET_INVALID_STATUS_CALLBACK): Drop 'L' modifier. 	* lmaccess.h (DEF_MIN_PWAGE): Ditto. 	* mmsystem.h (MEVT_EVENTPARM): Ditto. 	* svrapi.h (FRK_INIT):  Ditto. 	* vfw.h: Throughout, cast all constants in calls to SendMessage and 	ICSendMessage to the expected type. 	* prsht.h: Ditto, in calls to SendMessage and PostMessage. 	* winuser.h: Ditto, in calls to CreateWindowEx, CreateDialogParam, 	CreateDialogIndirectParam, DialogBoxParam, and DialogBoxIndirectParam. 	* commctrl.h: Ditto, in calls to SendMessage. 	(LPSTR_TEXTCALLBACKA): Drop 'L' modifier. 	(LPSTR_TEXTCALLBACKW): Ditto. 	* windowsx.h: Throughout, cast all constants in calls to SendMessage 	and as return value from HANDLE_xxx macro. 	(IsRestored): Replace test for == 0L with logical negation operator.
git-svn-id: svn+ssh://svn.code.sf.net/p/mingw-w64/code/trunk@5262 4407c894-4637-0410-b4f5-ada5f102cad1

+++ mingw-w64-headers/include/wmiatlprov.h
@@ -257,7 +257,7 @@ namespace ATL {
       }
       delete[] wClass;
       __MINGW_EXTENSION unsigned __int64 ullPathInfo;
-      hr = pPath->GetInfo(0L,&ullPathInfo);
+      hr = pPath->GetInfo((ULONG)0,&ullPathInfo);
       if(FAILED(hr)) return hr;
       if(!(ullPathInfo & ullTest)) return WBEM_E_INVALID_OBJECT_PATH;
       return WBEM_S_NO_ERROR;
